requirements.txt
discord.py==2.3.2
openai==1.3.2
python-dotenv==0.21.0
requests==2.25.1
